+++ app/Console/Commands/InventoryImporterCommand.php
@@ -54,27 +54,34 @@ class InventoryImporterCommand extends Command
 
         $networkHosts = NetworkHost::fetchHosts();
 
-        if ($networkHosts->count() === 0) {
+        $networkHostCount = $networkHosts->count();
+        if ($networkHostCount === 0) {
             $this->info('[Inventory Importer] Stopped importing. No hosts found.');
 
             return;
         }
 
-        try {
-            /** @var NetworkHost $networkHost */
-            foreach ($networkHosts as $networkHost) {
+        $networkHostSuccessfulConnectionCount = 0;
+        /** @var NetworkHost $networkHost */
+        foreach ($networkHosts as $networkHost) {
+            try {
                 $this->inventoryImporterService->import($networkHost->info->getInventoryUrl());
 
                 $networkHost->connectionSuccessful();
+                ++$networkHostSuccessfulConnectionCount;
+            } catch (UnexpectedClientResponseException $exception) {
+                $networkHost->connectionFailed();
+
+                Log::error(sprintf('[Inventory Importer] %s', $exception->getMessage()));
             }
-        } catch (UnexpectedClientResponseException $exception) {
-            $networkHost->connectionFailed();
-
-            Log::error(sprintf('[Inventory Importer] %s', $exception->getMessage()));
-
-            return;
         }
 
-        $this->info('[Inventory Importer] Finished importing data from all inventories.');
+        $this->info(
+            sprintf(
+                '[Inventory Importer] Finished importing data from %d/%d inventories.',
+                $networkHostSuccessfulConnectionCount,
+                $networkHostCount
+            )
+        );
     }
 }