+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2019-03-08
+### Changed
+- Inventory import - process all network hosts
+
 ## [0.3.1] - 2019-03-08
 ### Added 
 - Backward compatibility for Service Discovery 
@@ -23,8 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v0.3.1...master
-[0.3.0]: https://github.com/adshares/adserver/compare/v0.3.0...v0.3.1
+[Unreleased]: https://github.com/adshares/adserver/compare/v0.3.2...master
+[0.3.2]: https://github.com/adshares/adserver/compare/v0.3.1...v0.3.2
+[0.3.1]: https://github.com/adshares/adserver/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/adshares/adserver/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/adshares/adserver/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/adshares/adserver/compare/v0.1.0...v0.2.0